First step to make the webpage view a second class citizen.
