+++ src/App.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useRef } from "react";
 import "./App.css";
 import { SidebarLeft } from "@/components/sidebar-left";
 import { Separator } from "@/components/ui/separator";
@@ -7,11 +6,6 @@ import {
   SidebarProvider,
   SidebarTrigger,
 } from "@/components/ui/sidebar";
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "@/components/ui/resizable";
 import { AssistantChatContainer } from "@/components/ai-chat";
 import { SettingsProvider, SettingsView } from "@/components/settings";
 import { useAppStore } from "@/store/appStore";
@@ -22,10 +16,8 @@ import { InitializationError } from "@/components/InitializationError";
  * Inner app component that uses Zustand store
  */
 function AppContent() {
-  const containerRef = useRef<HTMLDivElement | null>(null);
   const {
     activeTaskId,
-    setContainerRef,
     showSettings,
     setShowSettings,
   } = useAppStore();
@@ -42,69 +34,35 @@ function AppContent() {
     return page?.url || null;
   });
 
-  /**
-   * Set the container ref in the store when it's available
-   */
-  useEffect(() => {
-    setContainerRef(containerRef);
-  }, [setContainerRef]);
-  
-  /**
-   * Force update bounds when container is mounted
-   */
-  useEffect(() => {
-    // Small delay to ensure DOM is fully rendered
-    const timer = setTimeout(() => {
-      if (containerRef.current) {
-        useAppStore.getState().updateContainerBounds();
-      }
-    }, 100);
-    
-    return () => clearTimeout(timer);
-  }, []);
 
   return (
     <div className="w-dvw flex flex-row h-dvh">
-      <ResizablePanelGroup direction="horizontal">
-        <ResizablePanel defaultSize={70}>
-          <SidebarProvider>
-            <SidebarLeft />
-            <SidebarInset className="relative">
-              <header className="bg-background sticky top-0 flex h-14 shrink-0 items-center gap-2">
-                <div className="flex flex-1 items-center gap-2 px-3">
-                  <SidebarTrigger />
-                  <Separator
-                    orientation="vertical"
-                    className="mr-2 data-[orientation=vertical]:h-4"
-                  />
-                  <div className="flex-1">
-                    {showSettings
-                      ? "Settings"
-                      : selectedPageUrl || "Project Management & Task Tracking"}
-                  </div>
-                </div>
-              </header>
-              <div
-                ref={containerRef}
-                className="relative flex flex-1 flex-col overflow-hidden h-full"
-              >
-                {showSettings && (
-                  <SettingsView onClose={() => setShowSettings(false)} />
-                )}
+      <SidebarProvider>
+        <SidebarLeft />
+        <SidebarInset className="relative flex-1">
+          <header className="bg-background sticky top-0 flex h-14 shrink-0 items-center gap-2">
+            <div className="flex flex-1 items-center gap-2 px-3">
+              <SidebarTrigger />
+              <Separator
+                orientation="vertical"
+                className="mr-2 data-[orientation=vertical]:h-4"
+              />
+              <div className="flex-1">
+                {showSettings
+                  ? "Settings"
+                  : selectedPageUrl || "Project Management & Task Tracking"}
               </div>
-            </SidebarInset>
-          </SidebarProvider>
-        </ResizablePanel>
-        <ResizableHandle />
-        <ResizablePanel
-          defaultSize={30}
-          minSize={15}
-          maxSize={75}
-          className="h-full overflow-hidden"
-        >
-          <AssistantChatContainer taskId={activeTaskId} />
-        </ResizablePanel>
-      </ResizablePanelGroup>
+            </div>
+          </header>
+          <div className="relative flex flex-1 flex-col overflow-hidden h-full">
+            {showSettings ? (
+              <SettingsView onClose={() => setShowSettings(false)} />
+            ) : (
+              <AssistantChatContainer taskId={activeTaskId} />
+            )}
+          </div>
+        </SidebarInset>
+      </SidebarProvider>
     </div>
   );
 }
